test-dspy/requirements.txt
dspy-ai>=2.0.0
pandas>=1.5.0
scikit-learn>=1.0.0
typing>=3.7.4
openai>=1.0.0 